Perform blacklisting based on reverse-host matching
The blocklists are purely host based, so it's more efficient to just
match on the host itself (although there is an, as of yet undetermined,
cost to cutting up the URL to obtain the host). This also works better
for some sites where resources might be loaded from the same host
as the site, but where the URL might contain other (blacklisted)
domains.  (This approach also saves us from iterating over the resource
URL which should compensate for the added getHost() call.)

E.g. nzz.ch will load image resources from nzz.ch, but the path
contains a blacklisted aws domain. We still want to load those images.
The old approach would perform a search on the entire URL, which would
result in those images being blocked.