Nullify callback to not receive unnecessary events after exiting browser (#171)
Callback events are meaningless once the fragment is no longer in use.
Our callback could potentially do bad stuff (e.g. try to access the
activity, or try to access strings via fragment utils, which depend
on being attached to an activity), so we should disable that possibility
to avoid crashes.