Update external-app dialog styling
Trying to set the style using:
<item name="alertDialogStyle">@style/DialogStyle</item>
in AppTheme, results in:
android.content.res.Resources$NotFoundException: Resource ID #0x0

We're passing in the correct activity, so I've got no idea why this is
happening. Explicitly setting the style when setting the dialog works
so lets just stick with that for now. (Extending AlertDialog.AppCompat
results in the theming being ignored. Even an empty style that
extends Theme.AppCompat.Dialog.Alert results in crashes unless we set
it in code, so it seems something is wrong with only the xml codepath.)