Don't use external URI handling for error: pages (#279)
I need to refactor our whole URL and error handling to make this cleaner.
This is complicated because sometimes URL are entered by the user
(in which case shouldOverrideUrlLoading() isn't called), and sometimes
they come from the webview (in which case we can override loading). We
want to make sure we get consistent behaviour in both cases.

(However we might want to exclude error: URLs when opening via a link?)