android2po: handle <strings> with multiple html children tags (#286)
Our previous implementation depended on there being only one html
tag child in the <string>, that was the case for most of our error
pages (<ul>, with further tags inside it), but our about page
is simpler (a list of <p>'s at the root, with other content too).
We therefore need to join those to get the full string representation
of <string> contents.