Bring URL input padding to spec (#299)
We only want 8dp between the edge of the URL input area, and the (X).
Similarly 8dp between the edge of the text, and the (X) or input area.
Since the (X) isn't always visible (i.e. when we're showing the hint
there's no (X)), it's easiest to set only right paddings, which guarantees
8dp in all correct locations and cases.