preparation for lint cleanup: don't use inexistant strings
These strings don't exist in the main strings.xml - they
only exist in the (outdate/unused/unupdated/historical) values-id/values-he
initial string imports (those were accidentally left in the repo,
and will be cleaned up).

Fortunately we still have a refresh string, but we don't have a back
string. Android doesn't provide one, and we're in a string freeze.
We need to readd this ASAP, see #353.