Update URL bar with real URL when loading finishes (#377)
We want to be able to show a sensible URL during loading, however
we might get redirected, in which case the new page URL won't
trigger a new request (try e.g. loading mozilla.org, which will
redirect to mozilla.org/en-US/ without a separate
onShouldInterceptRequest() for mozilla.org/en-US, in which
case we would keep displaying mozilla.org unless we check the
URL again once loading has finished).

Note: on iOS, the equivalent localContentBlocker() is called
with the redirected URL, which is why they're able to update
the URL bar much earlier.