Only load SearchEnginePreference title when actually displaying on-screen (#405)
At this point, search engine data will definitely have been loaded
(we load it in the background during app startup). SearchEnginePreference
is actually created on first-run (in the background) when setting
the default preferences, we don't want to block on waiting for
search engine data at that point, moreover we don't even need the data at
that point.