Enable strict mode after setting default preferences values.
Technically this is already our first strict mode violation - reading the preferences
will require reading data from the disk and we are on the main thread here. However
currently the following code requires that those values exist so we initialize
strict mode a little later.

Instead we could move the default values into code (constants) and avoid needing
to read the file here. We will have to keep them in sync though (code and xml).