No longer update currentPageURL in shouldOverrideUrlLoading()
This was needed when we used the pre-api21 version of shouldInterceptRequest.
shouldInterceptRequest() is called for the main document before
onPageStarted(). The api21 version of shouldInterceptRequest()
tells us when we're loading the main page, in which case we already
don't do any URL filtering - so there's no need for us to grab
the URL earlier.

shouldOverrideUrlLoading() is also called for iframes, which we definitely
don't want, since that could break whitelisting. (E.g. if you
visit iAmBlocked.foo which loads resources from iAmAlsoBlocked.foo, and
the entitylist allows iAmBlocked to load from iAmALsoBlocked - if
we then load an iframe with data:text/html then requests to iAmAlsoBlocked
will unintentionally break.)