UrlUtils: Scheme passed to isPermittedResourceProtocol() and isSupportedProtocol() can be null. (#899)
In various places we pass the value from Uri.getScheme() to those methods. However getScheme() can
return null and in those cases we crash. I haven't been able to actually find a way to trigger
this crash. But it seems to be rare looking at the number of crash reports we received.