Issue #1214: Convert IconGenerator to Kotlin.
The gains are that `getRepresentative*` are smaller and more readable.

One side effect is that I specified String? for the argument type because I
don't know enough about the code to know whether or not Java might be passing
a null value. I pass this down to the bottom-most method because it allowed me
to cleanly specify the default character ("?") only once.

I'm not sure if now is an appropriate time to convert this file, however.

This conversion was largerly automated and the tests still pass (yay, tests!)
which make me confident the conversion was a success.