Issue #1306: Do not erase WebView's Local Storage directory; fix test.
This fixes the user experienced issue where Google Search History is not erased.
However, the google search result page restores a google.com file in the "Local
Storage" directory so not all data on disk is actually deleted. I'll tackle this
in a later commit.

I verified that the correct files were deleted via WebViewDataTest.