Closes #1306: Delete local storage when starting a new browser session.
This code is hacky: see below.

We run clean-up on the main thread in order to guarantee we don't accidentally
delete current session data if we run asynchronously. I haven't seen any
problems with this locally but I'm running in an emulator. Alternative solutions
would be:
1. Find another place to delete (UrlInputFragment.onFilter worked before)
2. Run this on a background thread and run the fragment methods after (may need
to then run those on the UIThread). This would at least free the main thread
for UI events.

It creates more inconsistent confusion in our erase strategy. We should revise
this in #1472.