Closes #1842: Whitelist AS profiler files.
Given the conditions under which I've noticed these files are added, it's only
likely to occur in local builds. However, if we wanted to be extra safe that
these files aren't kept on disk in release builds, we could explicitly delete
them but then we'd presumably have to add a branch to allow developer builds to
use these files without problems. For simplicity, I've opted for the whitelist.