Closes #5020: Clear selected search suggetion from view model after processing.
The lifetime is bound to the lifetime of the activity. This is already questionable,
but currently two fragments communicate via this view model. If a search suggestion
is selected this is communicated via the view model. The next time we go to the
search screen the old selected search suggestion is still in the state and gets
processed again. This quick fix will take care of clearing the search suggestion
after processing it.

(cherry picked from commit faa22f07eb22909c9981fdf0b592383113db0e04)