For #5133: Added AppAction to skip onboarding
For perf testing, we need to be able to skip the onboarding screen. The easiest way could've been to change the share pref
but, in my opinion, we should try to keep things the same. Since everything is done through the Store to change the state of the app
I thought it would be better to have an action to put the app in 'perf test' mode